Added verification of MAC address
In the first octet of the node MAC address the RFC4122 specifies that the least significant bit is set if the MAC address is a multicast (bit set) that should be set only if the a random value is used. On the other hand, this lsb is not set if it is a unicast address (valid according to RFC4122). Based on experiments we can see that Windows not necessarily respect this making it harder to recognize non valid MAC addresses. Therefore we have added detection of the second least significant bit in the 1 octet of the node address. If it is set it means this is a locally assigned MAC address, which means it may not be globally unique. If either of these two bits are set, this may be an indication that the MAC address may not be valid.